VG19 Fouriertransform manuell getestet. Werte entsprechen Erwartungen.

+++ app/src/main/java/de/voicegym/voicegym/FourierHelper/PCMUtil.kt
@@ -4,8 +4,8 @@ object PCMUtil {
     fun getDoubleArrayFromShortArray(normalization: Double, inputArray: ShortArray): DoubleArray =
             inputArray.toList().map {
                 if (it >= 0)
-                    normalization * (it / Short.MAX_VALUE)
+                    normalization * (it.toDouble() / Short.MAX_VALUE)
                 else
-                    -normalization * (it / Short.MIN_VALUE)
+                    -normalization * (it.toDouble() / Short.MIN_VALUE)
             }.toDoubleArray()
 }
